Fix exception when using the testfile editor
The usage of depracted APIs caused exceptions to be thrown when the
outline view was used in combination with the testfile editor.

As far as I can tell, these APIs were depracted in Eclipse 3.0 and the
compatibility package has been removed in Eclipse 4.6. This caused
NullPointerExceptions to be thrown, essentially rendering the outline
view for RTS files useless.